# render: improve olm.bundle.object rendering for bundles