--- a/alpha/property/scheme.go
+++ b/alpha/property/scheme.go
@@ -12,6 +12,7 @@ func init() {
 		reflect.TypeOf(&GVK{}):             TypeGVK,
 		reflect.TypeOf(&GVKRequired{}):     TypeGVKRequired,
 		reflect.TypeOf(&BundleObject{}):    TypeBundleObject,
+		reflect.TypeOf(&CSVMetadata{}):     TypeCSVMetadata,
 		// NOTICE: The Channel properties are for internal use only.
 		//   DO NOT use it for any public-facing functionalities.
 		//   This API is in alpha stage and it is subject to change.