--- a/pkg/api/api_to_model.go
+++ b/pkg/api/api_to_model.go
@@ -5,6 +5,8 @@ import (
 	"fmt"
 	"sort"
 
+	"github.com/operator-framework/api/pkg/operators/v1alpha1"
+
 	"github.com/operator-framework/operator-registry/alpha/model"
 	"github.com/operator-framework/operator-registry/alpha/property"
 )
@@ -106,10 +108,13 @@ func convertAPIBundleToModelProperties(b *Bundle) ([]property.Property, error) {
 		out = append(out, property.MustBuildGVKRequired(p.Group, p.Version, p.Kind))
 	}
 
-	// If there is a bundle image reference, only create a bundle object property for the CSV.
-	// Otherwise, create a bundle object property for each object in the bundle.
-	if b.BundlePath != "" {
-		out = append(out, property.MustBuildBundleObjectData([]byte(b.CsvJson)))
+	// If there is a bundle image reference and a valid CSV, create an
+	// olm.csv.metadata property. Otherwise, create a bundle object property for
+	// each object in the bundle.
+	var csv v1alpha1.ClusterServiceVersion
+	csvErr := json.Unmarshal([]byte(b.CsvJson), &csv)
+	if csvErr == nil && b.BundlePath != "" {
+		out = append(out, property.MustBuildCSVMetadata(csv))
 	} else {
 		for _, obj := range b.Object {
 			out = append(out, property.MustBuildBundleObjectData([]byte(obj)))