--- a/pkg/registry/registry_to_model.go
+++ b/pkg/registry/registry_to_model.go
@@ -5,6 +5,8 @@ import (
 	"fmt"
 	"sort"
 
+	"github.com/operator-framework/api/pkg/operators"
+
 	"github.com/operator-framework/operator-registry/alpha/property"
 )
 
@@ -99,8 +101,13 @@ func ObjectsAndPropertiesFromBundle(b *Bundle) ([]string, []property.Property, e
 		if err != nil {
 			return nil, nil, fmt.Errorf("marshal object %s/%s (%s) to json: %v", obj.GetName(), obj.GetNamespace(), obj.GroupVersionKind(), err)
 		}
-		props = append(props, property.MustBuildBundleObjectData(objData))
 		objects = append(objects, string(objData))
+
+		// Only make an olm.bundle.object property if the object is a ClusterServiceVersion
+		// or if there is no bundle image set.
+		if obj.GetKind() == operators.ClusterServiceVersionKind || b.BundleImage == "" {
+			props = append(props, property.MustBuildBundleObjectData(objData))
+		}
 	}
 
 	if packageProvidedProperty == nil {