This pull request is a series of commits that progressively improves FBC parsing. Commits are:
1. Render `olm.bundle.object` _just_ for CSVs with `opm render <bundleImage>`
2. Render `olm.bundle.object` _just_ for CSVs with `opm render <sqliteImageOrFile>` (but only if the bundle has an image reference in the database).
3. Use $numCPUs goroutines to parse files in an FBC directory concurrently.
4. Introduce an `olm.csv.metadata` property, and use it in place of the `olm.bundle.object` properties in the scenarios covered in commits 1 and 2. When serving the GRPC API, this property is used to generate a synthetic CSV to be served.
5. Fix a bug in `Meta` object unmarshaling that was made apparent when testing commit 4 on real-world catalogs

A few notes:
- Commits 1 and 2 are somewhat overtaken by commit 4 (no code path generates full `olm.bundle.object` properties for _just_ CSVs after commit 4).
- Only commit 3 will realize gains for existing FBCs. Commits 1, 2, and 4 require re-rendering or re-migrating from bundles/sqlite catalogs (or manually pruning/converting `olm.bundle.object` properties)
- This PR does _not_ attempt to proactively remove objects from non-channel-heads. Despite the inability for existing OLMv0 components to make use of this metadata, it is nonetheless conceivably valuable for it to remain in catalogs to be readable by future versions of OLM.
- Commit 4's functionality is only activated if a bundle both:
    - has no CSV in any defined `olm.bundle.object` properties
    - has exactly 1 `olm.csv.metadata` property.
- Commit 4 injects the icon defined in `olm.package` into synthetic CSVs, if defined in the package.
- Commit 4 injects the description defined in `olm.package` into synthetic CSVs, if not defined in the `olm.csv.metadata` object.

Open Questions:
- How much of the `olm.csv.metadata` object belongs in the bundle vs. the package (e.g. display name, keywords, maintainers, provider)? In this PR, I'm hesitant to add new root level fields or package properties, but maybe that's an unfounded fear. It would be a further optimization because catalogs wouldn't have to repeat those same things over and over again in each bundle.
- Should we update the `render` and `migrate` commands to do any conversion of `olm.bundle.object` to `olm.csv.metadata` for existing FBCs? It seems like this would be pretty useful, so I've got another commit in a separate branch that would do this. I can include it in this PR or make a separate PR for it after we finalize this PR. Thoughts?
- This does PR does break an invariant we had originally tried to keep. That is "an sqlite db and its migrated FBC equivalent will serve the same GRPC API". Now that `opm migrate` intentionally loses fidelity, that invariant no longer holds. Are we okay with this? IMO, I think this is okay. SQLite serving has been deprecated for several years, so I think it is okay to start diverging. Perhaps we even consider dropping the SQLite server prior to the SQLite-based catalog building commands?

/hold
There's a bunch here to discuss, evaluate, and manually test. I updated our tests as best I could, and I think I've covered old code paths + new code paths. But this is a pretty significant change, so it would be good to get some detail-oriented eyes on this.

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>

**Reviewer Checklist**
- [ ] Implementation matches the proposed design, or proposal is updated to match implementation
- [x] Sufficient unit test coverage 
- [ ] Sufficient end-to-end test coverage
- [ ] Docs updated or added to `/docs` 
- [x] Commit messages sensible and descriptive


<!--

Note: If this PR is fixing an issue make sure to add a note saying:
Closes #<ISSUE_NUMBER>

-->