@joelanford 2023-05-18
I've uncovered and fixed a few more issues:
- The concurrency logic needed one more fix to prevent deadlocks. Now channels reads and writes can be preempted by the context being canceled
- In testing an FBC that uses `olm.csv.metadata`, I found that the `csv.metadata.labels` needed to be preserved.

With these changes, I:
- Pulled down the latest operatorhub catalog and migrated it to use `olm.csv.metadata` instead of `olm.bundle.object` properties, rebuilt it with an opm binary from this branch, and pushed to a personal quay repo.
- Created a kind cluster and installed OLM using `operator-sdk olm install` (this includes a catalog source for the real operatorhubio catalog)
- Created a catalog source pointing to my personal rebuilt operatorhub catalog
- Queried for all packagemanifests (once for the real catalog, once for my rebuilt catalog)
- Filtered out  `spec.metadata.creationTimestamp` and `spec.metadata.labels[catalog*]` which are not expected to be the same
- Performed a diff and noticed that the packagemanifests output for the rebuilt catalog contained an extra relatedImage for the bundle image
- Filtered out `status.channels[].currentCSVDesc.relatedImages`, performed another diff, which turned nothing up.

The last test I want to do is to put the `opm` binary from this branch into the existing operatorhubio catalog without changing any other contents, and verify that that image produces the same packagemanifests as the original. Once I verify that, I'll unhold.