@stevekuznetsov 2023-05-19
A Go benchmark would be sweet to capture this improvement.