@joelanford 2023-05-20
> JSON bits continue to break my brain

I _think_ I understand it much better now that I know that `sigs.k8s.io/yaml.Unmarshal` into `json.RawMessage` does this: https://github.com/kubernetes-sigs/yaml/issues/94