@joelanford 2023-05-25
I added a go benchmark in 4cfabfe, just before the changes that added concurrency. Then I:
- ran the benchmark on 4cfabfe
- ran the same exact benchmark on aef5a6b – the commit that added concurrency
- updated the benchmark to use `olm.csv.metadata` properties instead of `olm.bundle.object` properties in fc25695, and then ran it.

`benchstat` results below:
```
goos: darwin
goarch: arm64
pkg: github.com/operator-framework/operator-registry/alpha/declcfg
          │ 00-4cfabfe.txt │           01-aef5a6b.txt            │           02-fc25695.txt           │
          │     sec/op     │    sec/op     vs base               │   sec/op     vs base               │
LoadFS-10    10667.6m ± 0%   1355.1m ± 1%  -87.30% (p=0.002 n=6)   109.4m ± 1%  -98.97% (p=0.002 n=6)
```

TL;DR
- _Just_ concurrency gets us ~8x faster on my 8-core machine
- Using `olm.csv.metadata` and dropping `olm.bundle.object` gets another ~12x faster using the generated data
- So overall improvement with generated data on my machine is ~97x faster