@joelanford 2023-05-31
Re: the open questions from the PR description:
1. I chatted some with @tlwu2013. Both of us had come to the same conclusion: let's not over-complicate this PR with other optimizations related to moving more fields out of the bundle and into parent FBC blobs. We can handle that sort of thing separately if/when there is more clarity around it
2. I opened #1104
3. No discussion on this point yet. My opinion remains that we can break this invariant. I'm still interested in having a conversation about removing `opm registry serve` (but still keeping the other deprecated `opm` commands that write the database for now). Opened #1105