@grokspawn 2023-06-06
I think we have more conversation to explore the case-insensitivity of schema keys (FWIW, folks filed a bug with go/json about it, and they basically said "too late... Hyrum's law"), but I don't want to hold this heavy PR up because of that discussion.  We just need to not forgot about it.  

Waiting for the CI to finish to push in.