@AdrianVasiliu 2023-06-29
@joelanford For some reason, for the catalog of our operator, upgrading operator-registry from 1.27.1 to 1.28.0 and testing in OpenShift 4.12.7, I observe the following:
* The `CatalogSource` passes in `Ready` status as expected.
* But the operator doesn't show up in OperatorHub anymore.

Can this be a side-effect of the changes brought by this PR? We do observe (with `opm render`) large changes consistent with those involved here - just that we weren't expecting it would break our catalog. 
The notes in https://github.com/operator-framework/operator-registry/releases/tag/v1.28.0 don't seem to mention breaking changes.
All in one:
* Is it a bug, or is it an expected breaking change?
* If so, what adjustments should we bring to our catalog/bundle build to make it work with operator-registry 1.28.0?