@joelanford 2023-06-29
The changes require the opm server to be using 1.28 as well. Can you confirm the version of open used I. The catalog image to serve the new FBC?